fix(quant): TurboQuantPress must return input dtype (8B bf16 attention crash)
8B run exposed it: PolarQuant computes fp32, TurboQuantPress returned
fp32, but the bf16 model's attention needs bf16 keys -> 'query.dtype
BFloat16 key.dtype float' RuntimeError. Cast the reconstruction back to
x.dtype (BUGPress already did). Masked at 1B (fp32). Adds a bf16 test.

Also commit the 8B hybrid results: near-lossless at scale (base 6.88;
BUG+TQ r128/4b +0.20 at 0.161x, r256/4b +0.05 at 0.317x) -- BUG is far
more effective at 8B than 1B (~5x smaller ppl penalty), confirming the
U-amortization / scale hypothesis.

+++ src/kvdlra/press/turbo_press.py
@@ -88,7 +88,9 @@ class TurboQuantPress(BasePress):  # type: ignore[misc]
         pq = self._get_pq(h * d, x.device)
         codes, norm = pq.quantize(vecs)
         rec = pq.dequantize(codes, norm)
-        return rec.reshape(bsz, t, h, d).permute(0, 2, 1, 3)
+        # PolarQuant computes in fp32; cast back to x's storage dtype so keys match
+        # the model's dtype (e.g. bf16 at 8B -- else attention dtype-mismatches).
+        return rec.reshape(bsz, t, h, d).permute(0, 2, 1, 3).to(x.dtype)
 
     def compress(
         self,
